fix(diag-overlay): extract LogEntryViewModel to separate file
XamlC cannot resolve nested record type LogEntryViewModel from
DiagnosticLogViewModel. Moving it to its own file so the XAML
clr-namespace binding can find it.
